Enable viewing a single side-by-side image
Move the Activity classes to an activities package and create a new
image package containing an abstract StereoImage class and a very simple
concrete SideBySideImage class. Quickly hook these up so that the image
path passed to the View activity is displayed when the activity is
entered.

Also switch the display ImageViews to have black backgrounds to be less
distracting and make the ActionBar text in the View activity white so it
can still be read.