Enable using Java 1.8 features
This required updating the Android plugin for Gradle and Gradle itself
to allow the project to build targeting the Java 1.8 features.  Also had
to update the build SDK version as the previous wasn't supported by the
latest plugin.

Update the code in a couple of places to use the new features to give
cleaner code.