Set correct versionName when releasing on Actions
When GitHub Actions is triggered by publishing a release, the annotated
tag (for only that release) is replaced by the checkout action with a
lightweight tag (see https://github.com/actions/checkout/issues/290).

Add the --tags argument to the git describe command called by the build
so that it can generate the correct versionName in this situation.