Upgrade to AGP 7.1 and Gradle 7.2
Upgrade to Android Gradle Plugin 7.1.  This requires Gradle 7.2 or above
so upgrade this too.

The lint configuration has changed slightly for 7.2, so make slight
adjustments to cater for this while keeping the behaviour the same.

There is also a new lint check for redundant activity labels which
picked up that the label attribute on the View activity is redundant so
remove this.