Upgrade to AGP 7.2 and Gradle 7.3.3
Upgrade to Android Gradle Plugin 7.2.  This requires Gradle 7.3.3 or
above so upgrade this too.

The lint checks in AGP 7.2 were failing as the vector paths in the app
icon were more than 800 characters long.  Reduce the length of these to
less than the threshold by reducing the precision of the paths (this
doesn't show at the sizes that the icon appears) and optimising out some
unnecessary spaces.