Update Gradle to version 8.4
Android Gradle Plugin 8.3 requires Gradle 8.4.  Update the clean task to
avoid using a (newly) deprecated function.

Also bump compileSdk to 34 to fix a new lint warning and suppress new
'Discouraged API' warnings for the app not being resizable and
ViewActivity having fixed portrait rotation, which are both for good
reason in this case.